snow/requirements.txt
pandas
streamlit
sentence-transformers
umap-learn
hdbscan
scikit-learn
plotly
flask